backend/src/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-swagger-ui
pytest==8.2.0
flake8==7.0.0
Flask-CORS
Flask-JWT-Extended
python-dotenv
pytest-dependency